imp: updating SubAction button list to active and desactive elements
- creating base MapData to use in next feature
