+++ TableRPG/Assets/Demo/Scripts/Wall/WallBuilder.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using TableRPG;
 
 public class WallBuilder : MonoBehaviour
 {
