+++ TableRPG/Assets/Demo/Scripts/Wall/WallUIConfig.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.IO;
+using TableRPG;
 
 public class WallUIConfig : MonoBehaviour
 {
