+++ TableRPG/Assets/Prefabs/SubActionsMenu/SubActionMenuType1.prefab
@@ -1725,6 +1725,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 98e4e0ee23fa5064eb88bbc46c0859ed, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  subActionsButtons:
+    selectImages:
+    - {fileID: 8320747958434318986}
 --- !u!114 &7792488862156925422
 MonoBehaviour:
   m_ObjectHideFlags: 0
