+++ TableRPG/Assets/Prefabs/SubActionsMenu/SubActionMenuType2.prefab
@@ -1214,6 +1214,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cc9a18e12b68e7d4383c6c0c698a4d19, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  subActionsButtons:
+    selectImages:
+    - {fileID: 7521097191758038012}
 --- !u!114 &7792488862156925422
 MonoBehaviour:
   m_ObjectHideFlags: 0
