+++ TableRPG/Assets/Prefabs/SubActionsMenu/SubActionMenuWalls.prefab
@@ -873,6 +873,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 54d51f94a3f7eb64d92fafdba441705a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  subActionsButtons:
+    selectImages:
+    - {fileID: 6873889460390935179}
+    - {fileID: 3601936007596898074}
 --- !u!114 &7792488862156925422
 MonoBehaviour:
   m_ObjectHideFlags: 0
