+++ TableRPG/Assets/Scripts/Actions/SubActionsViewer/Type1SubActionViewer.cs
@@ -6,13 +6,13 @@ namespace TableRPG
 {
     public class Type1SubActionViewer : SubActionViewer
     {
-
         #region UI
         public void OnClickGeneric()
         {
+            this.subActionsButtons.SelectButton(0, true);
             Debug.Log("OnClickGeneric");
         }
         #endregion
-        
+
     }
 }
