+++ TableRPG/Assets/Scripts/Data.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b317f74e71c655d4bad4b9887f605bd6
+guid: b481ec2185be95241bb0f95baef6a82d
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
