+++ TableRPG/Assets/Scripts/Data/MapData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2f1694f7a7f2a74419bbb3c937272ff2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
