+++ TableRPG/Assets/Scripts/ScriptableObject.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fd22b305fbd7a0f4190b069778036e53
+guid: b317f74e71c655d4bad4b9887f605bd6
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
