+++ TableRPG/Assets/Scripts/ScriptableObject/WorldStateToColorDictionary.cs.meta
