+++ TableRPG/Assets/Scripts/SubActionsButtons.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2502305f06d1b0499fa7b7349b4c510
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
